requirements.txt
# Core dependencies
linkedin-api>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
feedparser>=6.0.0

# Data analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
scipy>=1.7.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.1
pytest-sugar>=0.9.4
pytest-html>=3.1.1
pytest-timeout>=1.4.2
pytest-asyncio>=0.16.0
pytest-xdist>=3.0.0
requests-mock>=1.9.0

# Logging and monitoring
structlog>=21.1.0
prometheus-client>=0.11.0 